+++ metapackages/auto-instrumentations-node/CHANGELOG.md
@@ -46,6 +46,12 @@
     * @opentelemetry/instrumentation-pino bumped from ^0.31.0 to ^0.32.0
     * @opentelemetry/instrumentation-redis-4 bumped from ^0.32.0 to ^0.33.0
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @opentelemetry/instrumentation-express bumped from ^0.31.1 to ^0.31.2
+
 ## [0.33.0](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/auto-instrumentations-node-v0.32.1...auto-instrumentations-node-v0.33.0) (2022-09-27)